+++ config/test.exs
@@ -9,3 +9,7 @@ config :ex_unit,
 config :commanded,
   dispatch_consistency_timeout: 100,
   event_store_adapter: Commanded.EventStore.Adapters.InMemory
+
+config :commanded, Commanded.Aggregates.LifespanAggregate,
+  snapshot_every: 2,
+  snapshot_version: 1