+++ test/aggregates/aggregate_lifespan_test.exs
@@ -1,87 +1,203 @@
 defmodule Commanded.Aggregates.AggregateLifespanTest do
   use Commanded.StorageCase
 
-  alias Commanded.Aggregates.{Aggregate, BankRouter}
-  alias Commanded.ExampleDomain.BankAccount
-
-  alias Commanded.ExampleDomain.BankAccount.Commands.{
-    CloseAccount,
-    OpenAccount,
-    DepositMoney,
-    WithdrawMoney
-  }
-
-  alias Commanded.ExampleDomain.BankAccount.Events.BankAccountOpened
+  alias Commanded.Aggregates.{DefaultLifespanRouter, LifespanAggregate, LifespanRouter}
+  alias Commanded.Aggregates.LifespanAggregate.{Command, Event}
   alias Commanded.EventStore
   alias Commanded.EventStore.RecordedEvent
   alias Commanded.Registration
 
-  describe "aggregate started" do
+  describe "aggregate lifespan" do
     setup do
       aggregate_uuid = UUID.uuid4()
 
       {:ok, ^aggregate_uuid} =
-        Commanded.Aggregates.Supervisor.open_aggregate(BankAccount, aggregate_uuid)
+        Commanded.Aggregates.Supervisor.open_aggregate(LifespanAggregate, aggregate_uuid)
 
-      pid = Registration.whereis_name({BankAccount, aggregate_uuid})
+      pid = Registration.whereis_name({LifespanAggregate, aggregate_uuid})
       ref = Process.monitor(pid)
+      reply_to = self() |> :erlang.pid_to_list()
 
-      %{aggregate_uuid: aggregate_uuid, pid: pid, ref: ref}
+      %{aggregate_uuid: aggregate_uuid, pid: pid, ref: ref, reply_to: reply_to}
     end
 
-    test "should shutdown after timeout", %{aggregate_uuid: aggregate_uuid, ref: ref} do
-      :ok = BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: 10})
+    test "should use default `after_event/1` lifespan when none specified'", %{
+      aggregate_uuid: aggregate_uuid,
+      ref: ref
+    } do
+      :ok = DefaultLifespanRouter.dispatch(%Command{uuid: aggregate_uuid, action: :event})
+
+      refute_receive {:DOWN, ^ref, :process, _, :normal}, 10
+    end
+
+    test "should use default `after_command/1` lifespan when none specified'", %{
+      aggregate_uuid: aggregate_uuid,
+      ref: ref
+    } do
+      :ok = DefaultLifespanRouter.dispatch(%Command{uuid: aggregate_uuid, action: :noop})
+
+      refute_receive {:DOWN, ^ref, :process, _, :normal}, 10
+    end
+
+    test "should use default `after_error/1` lifespan when none specified'", %{
+      aggregate_uuid: aggregate_uuid,
+      ref: ref
+    } do
+      {:error, {:failed, nil, nil}} =
+        DefaultLifespanRouter.dispatch(%Command{uuid: aggregate_uuid, action: :error})
+
+      refute_receive {:DOWN, ^ref, :process, _, :normal}, 10
+    end
+
+    test "should call `after_event/1` callback function", %{
+      aggregate_uuid: aggregate_uuid,
+      reply_to: reply_to
+    } do
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :event,
+          reply_to: reply_to,
+          lifespan: :stop
+        })
+
+      assert_receive :after_event
+      refute_received :after_command
+      refute_received :after_error
+    end
+
+    test "should shutdown after `after_event/1` timeout", %{
+      aggregate_uuid: aggregate_uuid,
+      ref: ref,
+      reply_to: reply_to
+    } do
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :event,
+          reply_to: reply_to,
+          lifespan: :stop
+        })
+
+      assert_receive {:DOWN, ^ref, :process, _, :normal}
+    end
+
+    test "should call `after_command/1` callback function when no domain events", %{
+      aggregate_uuid: aggregate_uuid,
+      reply_to: reply_to
+    } do
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :noop,
+          reply_to: reply_to,
+          lifespan: :stop
+        })
+
+      assert_receive :after_command
+      refute_received :after_event
+      refute_received :after_error
+    end
+
+    test "should shutdown after `after_command/1` timeout", %{
+      aggregate_uuid: aggregate_uuid,
+      ref: ref,
+      reply_to: reply_to
+    } do
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :noop,
+          reply_to: reply_to,
+          lifespan: :stop
+        })
+
+      assert_receive {:DOWN, ^ref, :process, _, :normal}
+    end
+
+    test "should call `after_error/1` callback function on error", %{
+      aggregate_uuid: aggregate_uuid,
+      reply_to: reply_to
+    } do
+      {:error, {:failed, _reply_to, :stop}} =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :error,
+          reply_to: reply_to,
+          lifespan: :stop
+        })
+
+      assert_receive :after_error
+      refute_received :after_event
+      refute_received :after_command
+    end
+
+    test "should shutdown after `after_error/1` timeout", %{
+      aggregate_uuid: aggregate_uuid,
+      ref: ref,
+      reply_to: reply_to
+    } do
+      {:error, {:failed, _reply_to, :stop}} =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :error,
+          reply_to: reply_to,
+          lifespan: :stop
+        })
 
       assert_receive {:DOWN, ^ref, :process, _, :normal}
     end
 
     test "should not shutdown if another command executed", %{
       aggregate_uuid: aggregate_uuid,
-      ref: ref
+      ref: ref,
+      reply_to: reply_to
     } do
-      :ok = BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: 10})
-      :ok = BankRouter.dispatch(%DepositMoney{account_number: aggregate_uuid, amount: 10})
+      command = %Command{
+        uuid: aggregate_uuid,
+        action: :noop,
+        reply_to: reply_to,
+        lifespan: 25
+      }
+
+      :ok = LifespanRouter.dispatch(command)
+      :ok = LifespanRouter.dispatch(command)
+
+      assert_receive :after_command
+      assert_receive :after_command
 
       refute_receive {:DOWN, ^ref, :process, _, :normal}, 25
       assert_receive {:DOWN, ^ref, :process, _, :normal}
-    end
-
-    test "should use default lifespan when it's not specified'", %{
-      aggregate_uuid: aggregate_uuid,
-      ref: ref
-    } do
-      :ok = BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: 10})
-      :ok = BankRouter.dispatch(%WithdrawMoney{account_number: aggregate_uuid, amount: 10})
-
-      refute_receive {:DOWN, ^ref, :process, _, :normal}, 100
     end
 
     test "should stop process when requested", %{
       aggregate_uuid: aggregate_uuid,
       ref: ref
     } do
-      :ok = BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: 10})
-      :ok = BankRouter.dispatch(%CloseAccount{account_number: aggregate_uuid})
+      :ok =
+        LifespanRouter.dispatch(%Command{uuid: aggregate_uuid, action: :noop, lifespan: :stop})
 
       assert_receive {:DOWN, ^ref, :process, _, :normal}
     end
 
-    test "should adhere to aggregate lifespan when taking snapshot after receiving published event",
+    test "should adhere to aggregate lifespan after taking snapshot",
          %{
            aggregate_uuid: aggregate_uuid,
-           pid: pid,
            ref: ref
          } do
-      :ok = BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: 10})
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :event,
+          lifespan: :infinity
+        })
 
-      events = aggregate_uuid |> EventStore.stream_forward() |> Enum.to_list()
-
-      assert Process.alive?(pid)
-
-      # publish events to aggregate before taking snapshot
-      send(pid, {:events, events})
-
-      :ok = Aggregate.take_snapshot(BankAccount, aggregate_uuid)
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :event,
+          lifespan: :stop
+        })
 
       assert_receive {:DOWN, ^ref, :process, _, :normal}
       assert {:ok, _snapshot} = EventStore.read_snapshot(aggregate_uuid)
@@ -93,15 +209,25 @@ defmodule Commanded.Aggregates.AggregateLifespanTest do
            pid: pid,
            ref: ref
          } do
-      :ok = BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: 10})
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :event,
+          lifespan: :infinity
+        })
 
-      events = aggregate_uuid |> EventStore.stream_forward() |> Enum.to_list()
-
-      :ok = Aggregate.take_snapshot(BankAccount, aggregate_uuid)
+      :ok =
+        LifespanRouter.dispatch(%Command{
+          uuid: aggregate_uuid,
+          action: :event,
+          lifespan: 1
+        })
 
       assert Process.alive?(pid)
 
-      # publish events to aggregate after taking snapshot
+      events = aggregate_uuid |> EventStore.stream_forward() |> Enum.to_list()
+
+      # Publish events to aggregate after taking snapshot
       send(pid, {:events, events})
 
       assert_receive {:DOWN, ^ref, :process, _, :normal}
@@ -117,7 +243,7 @@ defmodule Commanded.Aggregates.AggregateLifespanTest do
         %RecordedEvent{
           event_id: UUID.uuid4(),
           stream_version: 1,
-          data: %BankAccountOpened{account_number: aggregate_uuid, initial_balance: 1_000}
+          data: %Event{uuid: aggregate_uuid}
         }
       ]
 
@@ -126,47 +252,18 @@ defmodule Commanded.Aggregates.AggregateLifespanTest do
 
       refute_receive {:DOWN, ^ref, :process, ^pid, _}
     end
-
-    test "should adhere to aggregate lifespan after error", %{
-      aggregate_uuid: aggregate_uuid,
-      ref: ref
-    } do
-      ib = "clearly invalid"
-
-      {:error, :invalid_initial_balance} =
-        BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: ib})
-
-      refute_receive {:DOWN, ^ref, :process, _, :normal}, 29
-      assert_receive {:DOWN, ^ref, :process, _, :normal}, 35
-    end
-
-    test "should adhere to timeout when no event returned", %{
-      aggregate_uuid: aggregate_uuid,
-      ref: ref
-    } do
-      :ok =
-        BankRouter.dispatch(%OpenAccount{account_number: aggregate_uuid, initial_balance: 100})
-
-      :ok = BankRouter.dispatch(%CloseAccount{account_number: aggregate_uuid})
-      :ok = BankRouter.dispatch(%CloseAccount{account_number: aggregate_uuid})
-
-      assert_receive {:DOWN, ^ref, :process, _, :normal}
-    end
   end
 
-  describe "deprecated `after_command/1` callback" do
+  describe "compile time safety" do
     test "should fail to compile when missing `after_event/1` function" do
       assert_raise ArgumentError,
-                   "Aggregate lifespan `BankAccountLifespan` does not define a callback function: `after_event/1`",
+                   "Aggregate lifespan `BankAccountLifespan1` does not define a callback function: `after_event/1`",
                    fn ->
                      Code.eval_string("""
                        alias Commanded.ExampleDomain.BankAccount
                        alias Commanded.ExampleDomain.BankAccount.Commands.{OpenAccount, DepositMoney}
 
-                       defmodule BankAccountLifespan do
-                         def after_command(%OpenAccount{}), do: 5
-                         def after_command(%DepositMoney{}), do: 20
-                         def after_command(_), do: :infinity
+                       defmodule BankAccountLifespan1 do
                        end
 
                        defmodule BankRouter do
@@ -174,7 +271,31 @@ defmodule Commanded.Aggregates.AggregateLifespanTest do
 
                          dispatch [OpenAccount],
                            to: BankAccount,
-                           lifespan: BankAccountLifespan,
+                           lifespan: BankAccountLifespan1,
+                           identity: :account_number
+                       end
+                     """)
+                   end
+    end
+
+    test "should fail to compile when missing `after_command/1` function" do
+      assert_raise ArgumentError,
+                   "Aggregate lifespan `BankAccountLifespan2` does not define a callback function: `after_command/1`",
+                   fn ->
+                     Code.eval_string("""
+                       alias Commanded.ExampleDomain.BankAccount
+                       alias Commanded.ExampleDomain.BankAccount.Commands.{OpenAccount, DepositMoney}
+
+                       defmodule BankAccountLifespan2 do
+                         def after_event(_event), do: :stop
+                       end
+
+                       defmodule BankRouter do
+                         use Commanded.Commands.Router
+
+                         dispatch [OpenAccount],
+                           to: BankAccount,
+                           lifespan: BankAccountLifespan2,
                            identity: :account_number
                        end
                      """)