+++ test/aggregates/support/bank_router.ex
@@ -2,8 +2,6 @@ defmodule Commanded.Aggregates.BankRouter do
   @moduledoc false
   use Commanded.Commands.Router
 
-  alias Commanded.Aggregates.BankAccountLifespan
-
   alias Commanded.ExampleDomain.{
     BankAccount,
     OpenAccountHandler,
@@ -18,26 +16,10 @@ defmodule Commanded.Aggregates.BankRouter do
     WithdrawMoney
   }
 
-  dispatch OpenAccount,
-    to: OpenAccountHandler,
-    aggregate: BankAccount,
-    lifespan: BankAccountLifespan,
-    identity: :account_number
+  identify BankAccount, by: :account_number
 
-  dispatch DepositMoney,
-    to: DepositMoneyHandler,
-    aggregate: BankAccount,
-    lifespan: BankAccountLifespan,
-    identity: :account_number
-
-  dispatch WithdrawMoney,
-    to: WithdrawMoneyHandler,
-    aggregate: BankAccount,
-    identity: :account_number
-
-  dispatch CloseAccount,
-    to: OpenAccountHandler,
-    aggregate: BankAccount,
-    lifespan: BankAccountLifespan,
-    identity: :account_number
+  dispatch OpenAccount, to: OpenAccountHandler, aggregate: BankAccount
+  dispatch DepositMoney, to: DepositMoneyHandler, aggregate: BankAccount
+  dispatch WithdrawMoney, to: WithdrawMoneyHandler, aggregate: BankAccount
+  dispatch CloseAccount, to: OpenAccountHandler, aggregate: BankAccount
 end