+++ test/aggregates/support/lifespan/default_lifespan_router.ex
@@ -1,0 +1,8 @@
+defmodule Commanded.Aggregates.DefaultLifespanRouter do
+  use Commanded.Commands.Router
+
+  alias Commanded.Aggregates.LifespanAggregate
+  alias Commanded.Aggregates.LifespanAggregate.Command
+
+  dispatch [Command], to: LifespanAggregate, identity: :uuid
+end