+++ test/aggregates/support/lifespan/lifespan_aggregate.ex
@@ -1,0 +1,38 @@
+defmodule Commanded.Aggregates.LifespanAggregate do
+  @moduledoc false
+  defstruct [:uuid, :reply_to, :lifespan]
+
+  defmodule Command do
+    defstruct [:uuid, :reply_to, :action, :lifespan]
+  end
+
+  defmodule Event do
+    defstruct [:uuid, :reply_to, :lifespan]
+  end
+
+  alias Commanded.Aggregates.LifespanAggregate
+
+  def execute(%LifespanAggregate{}, %Command{action: :event} = command) do
+    %Command{uuid: uuid, reply_to: reply_to, lifespan: lifespan} = command
+
+    %Event{uuid: uuid, reply_to: reply_to, lifespan: lifespan}
+  end
+
+  def execute(%LifespanAggregate{}, %Command{action: :noop}) do
+    []
+  end
+
+  def execute(%LifespanAggregate{}, %Command{action: :error} = command) do
+    %Command{reply_to: reply_to, lifespan: lifespan} = command
+
+    {:error, {:failed, reply_to, lifespan}}
+  end
+
+  # State mutators
+
+  def apply(%LifespanAggregate{} = state, %Event{} = event) do
+    %Event{uuid: uuid, reply_to: reply_to, lifespan: lifespan} = event
+
+    %LifespanAggregate{state | uuid: uuid, reply_to: reply_to, lifespan: lifespan}
+  end
+end