+++ test/aggregates/support/lifespan/lifespan_router.ex
@@ -1,0 +1,9 @@
+defmodule Commanded.Aggregates.LifespanRouter do
+  use Commanded.Commands.Router
+
+  alias Commanded.Aggregates.Lifespan
+  alias Commanded.Aggregates.LifespanAggregate
+  alias Commanded.Aggregates.LifespanAggregate.Command
+
+  dispatch [Command], to: LifespanAggregate, identity: :uuid, lifespan: Lifespan
+end